code review fixes 1
