+++ Directory.Build.props
@@ -7,5 +7,8 @@
 		<Nullable>enable</Nullable>
 		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
 		<EnabLeNETAnalyzers>true</EnabLeNETAnalyzers>
+		<CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
+		<EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
+		<IsPackable>false</IsPackable>
 	</PropertyGroup>
 </Project>
