+++ GenericsHomework.Tests/VennTests.cs
@@ -4,61 +4,6 @@ namespace GenericsHomework.Tests;
 
 public class VennTests
 {
-    #region Circle Tests
-    [Fact]
-    public void Name_Set_Success()
-    {
-        Circle<int> circle = new("Jimmy");
-        Assert.Equal("Jimmy", circle.Name);
-    }
-
-    [Fact]
-    public void Name_NullValue_ThrowsArgumentNullException()
-    {
-        Assert.Throws<ArgumentNullException>(() => new Circle<string>(null!));
-    }
-
-    [Fact]
-    public void AddItem_NonNullValue_Success()
-    {
-        Circle<string> circle = new("Jimmy");
-        circle.AddItem("Jimbob");
-        Assert.Equal("Jimbob", circle.Items!.Element);
-    }
-
-    [Fact]
-    public void Items_Empty_Success()
-    {
-        Circle<int> circle = new("Jimmy");
-        Assert.Equal("", circle.GetItems());
-    }
-
-    [Fact]
-    public void Items_TwoItems_Success()
-    {
-        Circle<string> circle = new("Jimmy");
-        circle.AddItem("jimmy's son");
-        circle.AddItem("jimmy's daughter");
-        Assert.Equal("jimmy's son, jimmy's daughter", circle.GetItems());
-    }
-
-    [Fact]
-    public void ToStringOneItem_Success()
-    {
-        Circle<string> circle = new("Jimmy");
-        Assert.Equal("Jimmy: ", circle.ToString());
-    }
-
-    [Fact]
-    public void ToString_TwoItems_Success()
-    {
-        Circle<string> circle = new("Jimmy");
-        circle.AddItem("jimmy's son");
-        circle.AddItem("jimmy's daughter");
-        Assert.Equal("Jimmy: jimmy's son, jimmy's daughter", circle.ToString());
-    }
-    #endregion
-    #region VennDiagram Tests
     [Fact]
     public void ToString_OneEmptyCircle_Success()
     {
@@ -93,5 +38,4 @@ public class VennTests
         vennDiagram.AddCircle(circle2);
         Assert.Equal("{Jimmy's Dad: }, {Jimmy's Mom: }", vennDiagram.ToString());
     }
-    #endregion
 }
